Cloud-Function-Data-Ingest/requirements.txt
gbfs_client==0.1.8
google-cloud-pubsub==2.21.0
requests
pytz